requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
pandas>=2.0.0
openpyxl>=3.1.2
xlsxwriter>=3.2.0
Werkzeug>=2.3.7
SQLAlchemy>=2.0.16
reportlab>=4.0.0
Flask-Cors>=4.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
